No need to have ulongs here.

+++ src/SoundFingerprinting.Tests/Integration/SpectralImageDaoTest.cs
@@ -32,7 +32,7 @@
 
             Assert.AreEqual(3, spectralImageDao.GetSpectralImagesByTrackReference(trackReference).Count());
             var ids = spectralImageDao.GetSpectralImagesByTrackReference(trackReference)
-                    .Select(dto => (ulong)dto.SpectralImageReference.Id)
+                    .Select(dto => (uint)dto.SpectralImageReference.Id)
                     .ToList();
             CollectionAssert.AreEqual(Enumerable.Range(1, 3), ids);
         }
